actuation: DRobotAbstractActuation: Return motor position as double
This way we can also determine movements below 1 for learning.

+++ src/drobot/actuation/DRobotAbstractActuation.cpp
@@ -19,7 +19,8 @@ void DRobotAbstractActuation::setInitialPositions() {
 
 }
 
-int DRobotAbstractActuation::getMotorPosition(int index) {
+double DRobotAbstractActuation::getMotorPosition(int index)
+{
 	return _motors[index]->getMotorPosition();
 }
